+++ owl-back/src/main/java/com/project/owlback/codereview/model/CodeReview.java
@@ -1,19 +1,10 @@
 package com.project.owlback.codereview.model;
 
-import com.project.owlback.codereview.dto.CodeReviewItemDto;
-import com.project.owlback.codereview.repository.CodeReviewTagRepository;
 import jakarta.persistence.*;
-import jakarta.validation.constraints.NotNull;
 import lombok.Getter;
 import lombok.Setter;
-import org.hibernate.annotations.ColumnDefault;
-import org.hibernate.validator.constraints.Length;
-import org.springframework.beans.factory.annotation.Autowired;
 
-import java.sql.SQLException;
 import java.time.Instant;
-import java.util.ArrayList;
-import java.util.List;
 
 @Entity
 @Getter
@@ -23,7 +14,7 @@ public class CodeReview {
 
     @Id
     @Column(name = "code_review_id", nullable = false)
-    private Integer id;
+    private Long id;
 
     @Column(name = "version_count", nullable = false)
     private Integer versionCount;
